Restart after timeout implemented
I use the Meteostick over a long USB connection (Meteostick <-10m-> Raspberry Pi4) and keep having the problem that the Meteostick sporadically no longer sends any data. I switched an active USB over RJ45 Ethernet extender adaptor in between which, however, did not solve the problem. As a solution to the problem, I have adapted the meteostick.py file and implemented a reset if a timeout is recognized when receiving. Communication with the Meteostick works after this code change. The timeout reset action is entered in the raspberry system log.